bronze/requirements.txt
pandas
sqlalchemy
psycopg2
pyarrow
google-cloud-storage